refactor(client): extract interactive-examples action handler hook
